chore(lint): upgrade ESLint to v9 with flat config
Replace the legacy `.eslintrc.js` with `eslint.config.mjs` (flat config) and
bump to ESLint v9.36+. The primary motivation is that ESLint v8's built-in
dotfile-ignore treats every file under `.claude/worktrees/<name>/` as ignored,
breaking `npm run lint` for any agent running in a worktree. Flat config drops
that heuristic entirely.

Modeled on Expensify's already-shipped flat-config migration. Notable changes
beyond the config rewrite:

- `eslint-config-expensify` 2.x → 3.0.1 (now an ESM flat config). Required
  bumping `eslint-plugin-react-hooks` to ^7.1.1 (and the override) because
  v3 uses `reactHooks.configs.flat.recommended`. The v7 plugin also bundles
  the React Compiler rules, alongside the existing `eslint-plugin-react-compiler`.
- `@typescript-eslint/{eslint-plugin,parser}` 7.x → `typescript-eslint` ^8.20
  umbrella. v8 dropped the formatting/stylistic rules; `eslint-config-airbnb-typescript`
  still references the old names, so 16 of those are explicitly turned off so
  ESLint v9 can finalize the config. Prettier handles formatting anyway.
- Added `eslint-seatbelt` ^0.1.3 (plugin + processor + settings block). The
  baseline TSV is not generated in this commit — run
  `SEATBELT_INCREASE=ALL ./scripts/lint.sh` once and commit the resulting
  `eslint.seatbelt.tsv` to grandfather existing violations.
- Dropped `eslint-plugin-storybook` (no `.storybook` dir in the repo) and
  `eslint-plugin-prettier` (replaced by `eslint-config-prettier` applied as
  the last config entry).
- `scripts/lint.sh` adds `--concurrency=auto --no-warn-ignored` plus
  `--experimental-require-module` in `NODE_OPTIONS` so the CJS
  `eslint-plugin-rulesdir` can `require()` the ESM rule files shipped by
  `eslint-config-expensify` v3 on Node 20.x (no-op on 22.12+).
- `.github/workflows/lint.yml` cache key now hashes `eslint.config.mjs`,
  cache path moved to `node_modules/.cache/eslint`, and added a
  retry-without-cache fallback to dodge the well-known typescript-eslint
  stale-cache false positives.

Verified `./scripts/lint.sh src/App.tsx` from inside `.claude/worktrees/...`
returns exit 0 — the dotfile-ignore bug is gone.
