+++ .github/workflows/lint.yml
@@ -26,12 +26,20 @@ jobs:
       - name: Cache ESLint cache
         uses: actions/cache@v4
         with:
-          path: .eslintcache
-          key: ${{ runner.os }}-eslint-${{ hashFiles('package-lock.json', '.eslintrc.js') }}
+          path: node_modules/.cache/eslint
+          key: ${{ runner.os }}-eslint-${{ hashFiles('package-lock.json', 'eslint.config.mjs') }}
           restore-keys: |
             ${{ runner.os }}-eslint-
 
+      # ESLint's cache doesn't track cross-file TypeScript dependencies, which
+      # can cause stale errors. If lint fails, clear the cache and retry to
+      # rule out false positives.
       - name: Lint JavaScript and Typescript with ESLint
-        run: npm run lint
+        run: |
+          if ! npm run lint; then
+            echo "Lint failed, clearing cache and retrying..."
+            rm -rf node_modules/.cache/eslint
+            npm run lint
+          fi
         env:
           CI: true
