+++ package.json
@@ -20,7 +20,7 @@
     "lint": "./scripts/lint.sh",
     "lint-changed": "./scripts/lintChanged.sh",
     "lint-watch": "onchange '**/*.{js,jsx,ts,tsx,mjs,cjs}' -- ./scripts/lint.sh {{changed}}",
-    "lint:quiet": "eslint . --max-warnings=0 --quiet --cache --cache-location=.eslintcache",
+    "lint:quiet": "eslint . --max-warnings=0 --quiet --cache --cache-location=node_modules/.cache/eslint",
     "gh-actions-build": "./.github/scripts/buildActions.sh",
     "gh-actions-validate": "./.github/scripts/validateActionsAndWorkflows.sh",
     "merge": "./scripts/mergeAndPush.sh",
@@ -178,8 +178,7 @@
     "@types/seedrandom": "^3.0.8",
     "@types/semver": "^7.5.8",
     "@types/underscore": "^1.11.15",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
+    "@eslint/eslintrc": "^3.2.0",
     "@typescript/native-preview": "^7.0.0-dev.20260219.1",
     "@vercel/ncc": "0.38.1",
     "ajv-cli": "^5.0.0",
@@ -190,23 +189,25 @@
     "babel-plugin-transform-remove-console": "^6.9.4",
     "babel-preset-expo": "54.0.7",
     "baseline-browser-mapping": "^2.10.29",
-    "eslint": "^8.57.0",
+    "eslint": "^9.36.0",
     "eslint-config-airbnb-typescript": "^18.0.0",
-    "eslint-config-expensify": "^2.0.60",
+    "eslint-config-expensify": "3.0.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^27.6.1",
-    "eslint-plugin-jsdoc": "^48.0.2",
+    "eslint-plugin-jest": "^29.0.1",
+    "eslint-plugin-jsdoc": "^60.7.0",
     "eslint-plugin-jsx-a11y": "^6.5.9",
-    "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-react-compiler": "^19.1.0-rc.2",
-    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-react-hooks": "^7.1.1",
     "eslint-plugin-react-native": "^4.1.0",
     "eslint-plugin-react-native-a11y": "^3.3.0",
-    "eslint-plugin-storybook": "^0.6.15",
-    "eslint-plugin-testing-library": "^6.2.2",
-    "eslint-plugin-you-dont-need-lodash-underscore": "^6.13.0",
+    "eslint-plugin-rulesdir": "^0.2.2",
+    "eslint-plugin-testing-library": "^7.11.0",
+    "eslint-plugin-you-dont-need-lodash-underscore": "^6.14.0",
+    "eslint-seatbelt": "^0.1.3",
+    "globals": "^15.14.0",
+    "typescript-eslint": "^8.20.0",
     "firebase-admin": "^13.10.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "29.7.0",
@@ -234,7 +235,7 @@
   "overrides": {
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "eslint-plugin-react-hooks": "4.6.2"
+    "eslint-plugin-react-hooks": "^7.1.1"
   },
   "expo": {
     "autolinking": {
